Slightly different launch for batch block adjust extractions
