+++ spatial_cells/spatial_cell_gpu_kernels.hpp
